+++ gcip/addons/container/registries.py
@@ -1,0 +1,7 @@
+class Registry():
+    """
+    Container registry urls constants.
+    """
+    DOCKER: str = "https://index.docker.io/v1/"
+    QUAY: str = "quay.io"
+    GCR: str = "gcr.io"